fix(schema): timeout and retry data type fixed
In 'ExtractionJobToLLM' schema the 'llm_timout' & 'llm_retry' were stored as 'float' data type
Previoustly when llm_extraction was hardcoded to Gemini the float was manually converted to int before calling Gemini_API

taking these params as float and conveting to int to pass to Instructor+LiteLLM takes extra step and makes code base confusing
So, initially validating it to int makes is much easier and no manually handling is required

+++ graphcrawl_ai/models/crawl_url/request_models/request_to_llm.py
@@ -32,11 +32,11 @@ class ExtractionJobToLLM(BaseModel):
         ...,
         description="The template model used to shape and format the AI's answer."
     )
-    llm_timeout: Optional[float] = Field(
+    llm_timeout: Optional[int] = Field(
         60, 
         description="How many seconds to wait for the AI to answer."
     )
-    llm_retry: Optional[float] = Field(
+    llm_retry: Optional[int] = Field(
         3, 
         description="How many times to try asking the AI again if it fails."
     )
